Set default options values correctly using function 'setDefaultOption', added options for slug change and post deletion automatic cache purge, improved tests

+++ src/Servebolt/FullPageCache/FullPageCache.php
@@ -4,6 +4,7 @@ namespace Servebolt\Optimizer\FullPageCache;
 
 if (!defined('ABSPATH')) exit; // Exit if accessed directly
 
+use Servebolt\Optimizer\CachePurge\CachePurge;
 use Servebolt\Optimizer\CachePurge\WordPressCachePurge\WordPressCachePurge;
 use Servebolt\Optimizer\Traits\Singleton;
 use function Servebolt\Optimizer\Helpers\checkboxIsChecked;
@@ -32,6 +33,16 @@ class FullPageCache
      */
     private function purgePostCacheIfAddedToFpcExclusion(): void
     {
+        // Skip this feature if the cache purge feature is inactive or insufficiently configured, or it automatic cache purge is inactive
+        if (!CachePurge::featureIsAvailable() || !CachePurge::automaticCachePurgeOnContentUpdateIsActive()) {
+            return;
+        }
+
+        // Should skip automatic cache purge for posts added to FPC exclusion?
+        if (apply_filters('sb_optimizer_disable_automatic_purge_on_post_added_to_fpc_exclusion', false)) {
+            return;
+        }
+
         add_action('sb_optimizer_post_added_to_fpc_exclusion', function($postId) {
             try {
                 WordPressCachePurge::purgeByPostId($postId);
